# docs/api/recovery-control-plane-openapi.yaml
openapi: 3.1.0
...
info:
  title: Windows Recovery MVP Control Plane API
  version: 1.0.0
  summary: Recovery MVP v1 的单实例 Control Plane 权威线上合同
  description: |
    Control Plane 接收 Agent 主动注册/心跳，维护服务镜像，并按严格依赖 DAG 调度
    RecoveryRun。Agent ingress 只接受实验 `cluster_token`；管理 API 只接受管理员签名
    Session，且所有管理写请求同时要求 CSRF Header。两类凭据不可互换。

    HTTP 明文模式只允许实验或受控管理网。Control Plane 必须部署在独立管理节点，
    其 SQLite 不得依赖被管理的 MySQL、Redis 或 Nacos。
servers:
  - url: http://{control_plane_ip}:{control_plane_port}
    description: 单实例 Windows Control Plane
    variables:
      control_plane_ip:
        default: 127.0.0.1
      control_plane_port:
        default: "8766"
...
security:
  - adminSession: []
paths:
...
  /api/v1/services:
    get:
      tags: [Inventory]
      operationId: listManagedServices
      summary: 列出 CP 服务镜像
      responses:
        "200":
          description: 每项以 CP 分配的 `managed_service_id` 标识。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedServiceCollection"
              example:
                items:
                  - managed_service_id: 72961a6e-35cd-46c8-b405-56d73dbf3619
                    agent_id: b8d82b70-5a8a-4c07-8f44-31ef85a6aa01
                    agent_online_status: ONLINE
                    local_service_id: mysql
                    windows_service_name: MySQL80
                    display_name: MySQL 8
                    installation_state: INSTALLED
                    runtime_state: ACTIVE
                    startup_state: AUTOSTART_DISABLED
                    last_observed_at: "2026-07-16T08:00:00Z"
                    mirrored_at: "2026-07-16T08:00:01Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/services/{managed_service_id}/actions/{action}:
    post:
      tags: [Service actions]
      operationId: proxyServiceAction
      summary: 代理单服务 start、stop 或 restart
      description: |
        CP 将 managed_service_id 映射为所属 Agent 的 local_service_id。涉及非终态 RecoveryRun
        独占的服务返回 `409 SERVICE_IN_ACTIVE_RUN`。Idempotency-Key 原样传给 Agent。正文可省略；
        若提供只能是严格空对象 `{}`。
      security:
        - adminSession: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/ManagedServiceId"
        - $ref: "#/components/parameters/ServiceAction"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmptyActionRequest"
            example: {}
      responses:
        "202":
          description: Agent Operation 已创建或命中幂等记录。
          headers:
            Location:
              $ref: "#/components/headers/OperationLocation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxiedOperation"
              example:
                managed_service_id: 72961a6e-35cd-46c8-b405-56d73dbf3619
                operation_id: 1e881030-2451-4d7f-9628-6a58c2912e14
                agent_id: b8d82b70-5a8a-4c07-8f44-31ef85a6aa01
                local_service_id: mysql
                windows_service_name: MySQL80
                action: start
                status: PENDING
                idempotency_key: db7df82f-c129-4e7a-bc51-6bcd4433fe74
                request_fingerprint: 7bc1e2673f2333ac92a479a3aa55e91fc19451269d5c2f33c3423178b00026e8
                created_at: "2026-07-16T08:00:02Z"
                updated_at: "2026-07-16T08:00:02Z"
                started_at: null
                finished_at: null
                error_code: null
                message: null
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/operations/{operation_id}:
    get:
      tags: [Service actions]
      operationId: getProxiedOperation
      summary: 查询 CP 代理创建的 Agent Operation
      parameters:
        - $ref: "#/components/parameters/OperationId"
      responses:
        "200":
          description: Operation 当前状态；UNKNOWN 不得自动重放。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProxiedOperation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    clusterToken:
      type: http
      scheme: bearer
      bearerFormat: opaque-cluster-token
      description: 仅用于 Agent 注册/心跳；至少 32 个随机字节并常量时间比较。
    adminSession:
      type: apiKey
      in: cookie
      name: recovery_admin_session
      description: 单管理员登录后签发的 HttpOnly、SameSite 签名 Session Cookie。
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: 所有管理 API 写请求必须与管理员 Session 中的 CSRF 值匹配。
  parameters:
...
    ManagedServiceId:
      name: managed_service_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UuidV4Input"
...
    OperationId:
      name: operation_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UuidV4Input"
    ServiceAction:
      name: action
      in: path
      required: true
      schema:
        type: string
        enum: [start, stop, restart]
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        $ref: "#/components/schemas/UuidV4Input"
...
  headers:
    OperationLocation:
      description: CP Operation 查询路径。
      required: true
      schema:
        type: string
        format: uri-reference
...
  responses:
    Unauthorized:
      description: 对当前接口缺少正确的 cluster token 或管理员 Session。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: AUTH_INVALID
            message: Authentication credential is missing or invalid
            detail: null
            request_id: 7cf653dc-3bba-428e-8d5e-1bfd9220a60f
    Forbidden:
      description: 管理写请求缺少/错误 CSRF，或来源策略拒绝。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: SOURCE_IP_DENIED
            message: Source address is not allowed
            detail: null
            request_id: 7cf653dc-3bba-428e-8d5e-1bfd9220a60f
    NotFound:
      description: Agent、service、Operation、group 或 Run 不存在。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: RUN_NOT_FOUND
            message: RecoveryRun was not found
            detail: null
            request_id: 7cf653dc-3bba-428e-8d5e-1bfd9220a60f
    Conflict:
      description: fencing、Agent 离线、依赖、组前置条件、幂等或服务独占冲突。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: GROUP_NOT_READY
            message: Recovery group preconditions are not satisfied
            detail:
              group_state: BLOCKED_PRECONDITION
            request_id: 7cf653dc-3bba-428e-8d5e-1bfd9220a60f
    UnprocessableEntity:
      description: JSON 可解析，但字段或跨字段约束无效。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: VALIDATION_ERROR
            message: One or more fields are invalid
            detail: null
            request_id: 7cf653dc-3bba-428e-8d5e-1bfd9220a60f
  schemas:
    UuidV4:
      type: string
      format: uuid
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
    UuidV4Input:
      type: string
      format: uuid
      pattern: "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-4[0-9A-Fa-f]{3}-[89ABab][0-9A-Fa-f]{3}-[0-9A-Fa-f]{12}$"
    Sha256:
      type: string
      pattern: "^[0-9a-f]{64}$"
    ServiceSlug:
      type: string
      minLength: 1
      maxLength: 64
      pattern: "^[a-z][a-z0-9]*(?:-[a-z0-9]+)*$"
    InstallationState:
      type: string
      enum: [INSTALLED, NOT_INSTALLED, UNKNOWN]
    RuntimeState:
      type: string
      enum: [ACTIVE, INACTIVE, STARTING, STOPPING, FAILED, UNKNOWN]
    StartupState:
      type: string
      enum: [AUTOSTART_ENABLED, AUTOSTART_DISABLED, START_BLOCKED, UNKNOWN]
    AgentOnlineStatus:
      type: string
      enum: [ONLINE, OFFLINE]
    OperationStatus:
      type: string
      enum: [PENDING, RUNNING, SUCCEEDED, FAILED, REJECTED, UNKNOWN]
...
    ErrorCode:
      type: string
      enum:
        - AUTH_REQUIRED
        - AUTH_INVALID
        - SOURCE_IP_DENIED
        - STALE_AGENT_INSTANCE
        - ENDPOINT_SOURCE_MISMATCH
        - ROUTE_NOT_FOUND
        - METHOD_NOT_ALLOWED
        - INTERNAL_ERROR
        - VALIDATION_ERROR
        - IDEMPOTENCY_KEY_REQUIRED
        - IDEMPOTENCY_KEY_INVALID
        - IDEMPOTENCY_KEY_REUSED
        - AGENT_NOT_FOUND
        - AGENT_OFFLINE
        - SERVICE_NOT_ALLOWLISTED
        - SERVICE_MAPPING_CHANGED
        - SERVICE_NOT_INSTALLED
        - SERVICE_STATE_UNKNOWN
        - SERVICE_ACTION_CONFLICT
        - SERVICE_IN_ACTIVE_RUN
        - SCM_ACTION_FAILED
        - SCM_ACTION_TIMEOUT
        - OPERATION_NOT_FOUND
        - AGENT_PROTOCOL_MISMATCH
        - PROBE_TARGET_DENIED
        - PROBE_UNSUPPORTED
        - PROBE_FAILED
        - DEPENDENCY_CYCLE
        - GROUP_NOT_READY
        - RUN_NOT_FOUND
    EmptyActionRequest:
      type: object
      maxProperties: 0
      additionalProperties: false
...
    ManagedService:
      type: object
      additionalProperties: false
      required: [managed_service_id, agent_id, agent_online_status, local_service_id, windows_service_name, display_name, installation_state, runtime_state, startup_state, last_observed_at, mirrored_at]
      properties:
        managed_service_id:
          $ref: "#/components/schemas/UuidV4"
        agent_id:
          $ref: "#/components/schemas/UuidV4"
        agent_online_status:
          $ref: "#/components/schemas/AgentOnlineStatus"
        local_service_id:
          $ref: "#/components/schemas/ServiceSlug"
        windows_service_name:
          type: string
          minLength: 1
          maxLength: 256
        display_name:
          type: string
          minLength: 1
          maxLength: 256
        installation_state:
          $ref: "#/components/schemas/InstallationState"
        runtime_state:
          $ref: "#/components/schemas/RuntimeState"
        startup_state:
          $ref: "#/components/schemas/StartupState"
        last_observed_at:
          type: string
          format: date-time
        mirrored_at:
          type: string
          format: date-time
    ManagedServiceCollection:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 1024
          items:
            $ref: "#/components/schemas/ManagedService"
    ProxiedOperation:
      type: object
      additionalProperties: false
      required: [managed_service_id, operation_id, agent_id, local_service_id, windows_service_name, action, status, idempotency_key, request_fingerprint, created_at, updated_at, started_at, finished_at, error_code, message]
      properties:
        managed_service_id:
          $ref: "#/components/schemas/UuidV4"
        operation_id:
          $ref: "#/components/schemas/UuidV4"
        agent_id:
          $ref: "#/components/schemas/UuidV4"
        local_service_id:
          $ref: "#/components/schemas/ServiceSlug"
        windows_service_name:
          type: string
          minLength: 1
          maxLength: 256
        action:
          type: string
          enum: [start, stop, restart]
        status:
          $ref: "#/components/schemas/OperationStatus"
        idempotency_key:
          $ref: "#/components/schemas/UuidV4"
        request_fingerprint:
          $ref: "#/components/schemas/Sha256"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        started_at:
          type: [string, "null"]
          format: date-time
        finished_at:
          type: [string, "null"]
          format: date-time
        error_code:
          oneOf:
            - $ref: "#/components/schemas/ErrorCode"
            - type: "null"
        message:
          type: [string, "null"]
          maxLength: 1024
...
    ErrorDetail:
      oneOf:
        - type: object
          additionalProperties: true
        - type: "null"
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message, detail, request_id]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          minLength: 1
          maxLength: 1024
        detail:
          $ref: "#/components/schemas/ErrorDetail"
        request_id:
          $ref: "#/components/schemas/UuidV4"